Update patch set 6
Abandoned

I do not believe we need this one: generated classes are part of implementation contract, so we do not expect them to be reflected over. This patch brings to the table only better reflection accuracy. Might be useful for debugging in future.

Patch-set: 6
Status: abandoned
Tag: autogenerated:gerrit:abandon