Map system-ordered keyed lists to Map
The type mapping for YANG lists which have a key and do not have
a user-specified order should point to java.util.Map, as that
provides the correct semantics.

This patch changes that mapping, resulting in better ergonomics,
as it is now possible to acquire specific items based on their
keys.

JIRA: MDSAL-434
Change-Id: If2f8980fde1c419bbc0b7c414d7deab329362b47
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>