WIP: Add mdsal-{dom,binding}-testkit(-spi)
Rather than providing abstract classes for subclassing by user
tests, we introduce TestKit classes, which provide MD-SAL services
with proper interfaces to control data flows.

This allows downstreams to properly dispatch events, so that some
of the uncertainities (mostly around DataTreeChangeListener) can
be properly asserted.

The functionality is split into three parts:
- mdsal-dom-testkit-spi, which provides the core DOM testkit
  functionality, without being dependent on mdsal-dom-broker.
  This allows clear implementation separation and reuse in
  mdsal-dom-broker tests.
- mdsal-dom-testkit, which hooks up mdsal-dom-testkit-spi
  with mdsal-dom-broker -- forming the functionality we'd want
  downstreams to ues
- mdsal-binding-testkit-spi, which serves the same purpose as
  its DOM counterpart, except it avoid the dependency on
  mdsal-binding-dom-adapter.

The fourth part already exists in the form of mdsal-binding-test-utils,
which are fitted with the binding testkit specialization.

JIRA: MDSAL-393
Change-Id: If6f979b0d1303f2efaaece4a954cc93d3da91400
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>