Return Ipv{4,6}AddressNoZone when formatting from binary
Ever since MDSAL-440 was implemented, there is no point in our
utilities returning Ipv4Address objects as they never contain a zone
and therefore are equivalent to Ipv4AddressNoZone.

Ipv4AddressNoZone has the additional benefit that the lack of a zone
is contained in the type definition -- hence it can be propagated
through Java interfaces in a type-safe manner. Futhermore, given
the relationship between the two classes, the non-presence of zone
can be ascertained via a simple instanceof check rather than
searching the string.

JIRA: MDSAL-511
Change-Id: I7e0ed85bad7a63af9e9a6f8838609c0e53776ab0
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>