Remove BindingRuntimeContext.getEnumMapping()
These methods are inexact and unused, remove them to prevent any
confusion as to what goes on w.r.t. enums: their class-to-YANG
mapping is captured in the classes themselves, so go fetch it as
appropriate: either enumerate the class, or if you have an instance,
just call getName().

JIRA: MDSAL-392
Change-Id: I0034004e92ae4444efab2401dc16b859047d677a
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>