WIP: Add simple query language
Applications want to perform rudimentary queries without having to
perform the filtering by themselves.

On the Binding layer this means providing enough expressive power
to formulate such queries in a type-safe manner and translating
them to their DOM counterparts. This has the benefit of allowing
filtering to occur in DOM world, side-stepping the need to instantiate
Binding objects.

On the DOM layer the expression can be transmitted, hence it gives
the possibility to move the execution to storage backend, thus
reducing app/backend data interchange volume.

// FIXME: finish DOM builder/query
// FIXME: finish Binding implementation
// FIXME: add tests for translation

Change-Id: I51c8ec7e34c2485f62aeb5bdbe35fe1507cabaa9
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>