Expose completion future from WriteOperations
WriteOperations (and its counterparts) does not allow code to control
when a transaction is committed or aborted. This adds a layer of
separation, but unfortunately also prevents the encapsulated code from
reacting to when the changes are actually committed (or not).

This capability is quite important for chaining cache updates and
similar tasks.

Expose a FluentFuture<?>. which is guaranteed to complete when its
transaction is completed -- either successfully or not.

JIRA: MDSAL-61
Change-Id: Ie75671842b93fb9e63f1c2aa9ec72f25904da039
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>