Correct ClassTemplate's check for types
ClassTemplate's checks run against Java Types, which is fine, except
that we should be referencing them through BaseYangTypes, so that we
have the incidental alignment well covered.

Furthermore the checks rely on FQCN, which is potentially slow to
compute, while the new checks rely on plain type equality -- hence
potentially speeding up things.

We also drop a FIXMEs to clarify situation around InstanceIdentifier.

Change-Id: I3328199a8ec56e28d58582c076f1344fb4459103
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>