Drop mdsal-binding-dom-adapter's dependency on in-memory datastore
We do not want to be pulling in in-memory datastore just for the
sake of making test-jar work. Now that we binding-test-utils, which
downstreams can nicely depend on, we can express the test-jar
dependency in a reasonable way.

This has the effect of removing the mdsal-dom-inmemory-datastore
from odl-mdsal-binding-dom-adapter.

JIRA: MDSAL-559
Change-Id: I723285a77e0d48ca36cce4d21fffbbcf2c2e9e25
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>