Improve the reconnecting and keepalive mechanisms
In some cases during a network partition the disconnected
channel gets closed with delay and after a new channel
is already created. This starts reconnection process
which closes the new channel and creates a another one.
Also the keepalive mechanism needs to be improved since
some types of network partition leaves one side unaware
of it. It is important that both the Sink and the Source
are notified about any connection issue ASAP.

Change-Id: Iebde72963bddb748ab97617d07cfc77cd8614da4
Signed-off-by: Tibor Král <tibor.kral@pantheon.tech>