Netty Replicator - improve the reconnection and keepalive mechanisms
- In some cases during a network partition the disconnected
channel got closed with delay after a new channel
was already created. This started reconnection process
which closed the new channel and created yet another one.
- Also improved the keepalive mechanism since some types
of network partitions left one side unaware of the issue.
It is important that both the Sink and the Source are
notified about any connection issue ASAP. PING-PONG
messages are exchanged between the two sides when no
deltas are sent for a period of time.
- config options keepalive-interval-seconds and
max-missed-keepalives added to both configurations.

Change-Id: Iebde72963bddb748ab97617d07cfc77cd8614da4
Signed-off-by: Tibor Král <tibor.kral@pantheon.tech>