Add non-null getters for leaf objects
We do not have a reasonable type to return for leaf types, as there is
no content we could ever garner - hence we cannot use 'nonnull'
prefix. For these we need a suitable prefix, which is TBD, and those
methods should be failing if the element does not exist.

JIRA: MDSAL-602
Change-Id: Ia66ee61f51fb9b6eb11aea3736e5246aadad6a3e
Signed-off-by: Kostiantyn Nosach <kostiantyn.nosach@pantheon.tech>